Change type date of int to double parameters

+++ src/main/java/net/thumbtack/denisenko/trainee/figure/geometry/Figure.java
@@ -15,7 +15,7 @@ public abstract class Figure implements Colored, Serializable{
         this.color = color;
     }
 
-    public abstract void move(int newX, int newY);
+    public abstract void move(double newX, double newY);
     public abstract double area();
     public abstract String printCoordinates();
     public abstract boolean isPointInFigure(Point2D point);
